move sensors to folder
